Implementation of optimal orientation weighted output
The branch ORIENTATION (-OR) is to implement the exploitation of the optimal
orientations found by the filter for each pixel. This will be (vectorially)
space-averaged around each pixel and then the modulus of this average will
be used to weight the filter response. This could give a further selectivity
and improve the enhancement of tubular/linear structures.